toolkit/components/url-classifier/tests/mochitest/test_classifier.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the URI Classifier</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var Cc = Components.classes;
var Ci = Components.interfaces;

// Add some URLs to the malware database.
var testData = "malware.com/"
var testUpdate =
  "n:1000\ni:test-malware-simple\nad:1\n" +
  "a:524:32:" + testData.length + "\n" +
  testData;

var dbService = Cc["@mozilla.org/url-classifier/dbservice;1"]
                .getService(Ci.nsIUrlClassifierDBService);

var numTries = 10;

function doUpdate(update) {
  var listener = {
    QueryInterface: function(iid)
    {
      if (iid.equals(Ci.nsISupports) ||
          iid.equals(Ci.nsIUrlClassifierUpdateObserver))
        return this;
      throw Cr.NS_ERROR_NO_INTERFACE;
    },

    updateUrlRequested: function(url) { },
    streamFinished: function(status) { },
    updateError: function(errorCode) {
      ok(false, "Couldn't update classifier.");
      SimpleTest.finish();
    },
    updateSuccess: function(requestedTimeout) {
      document.getElementById("testFrame").src = "classifierFrame.html";
    }
  };

  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  try {
    dbService.beginUpdate(listener,
                          "test-malware-simple", "");
    dbService.beginStream("", "");
    dbService.updateStream(update);
    dbService.finishStream();
    dbService.finishUpdate();
  } catch(ex) {
    // The DB service might already be updating.  Try again after a 5 seconds...
    if (--numTries != 0) {
      setTimeout(function() { doUpdate(update) }, 5000);
      return;
    }
    throw ex;
  }
}

doUpdate(testUpdate);

SimpleTest.waitForExplicitFinish();

</script>
</pre>
<iframe id="testFrame" onload=""></iframe>
</body>
</html>
